asdsad

+++ src/routes.js
@@ -1,6 +1,7 @@
 import { randomUUID } from 'node:crypto';
 import { buildRoutePath } from './utils/build-route-path.js';
-import { Database } from './database.js';
+import { Database, databasePath } from './database.js';
+import fs from 'fs/promises';
 
 const database = new Database;
 
@@ -9,15 +10,26 @@ export const routes = [
         method: 'POST',
         path: buildRoutePath('/tasks'),
         handler: (req, res) => {
+            const { title, description } = req.body;
 
-            database.insert('tasks', {
-                id: randomUUID(),
-                ...req.body
-            });
+            if(title && description) {
 
-            return res
+                database.insert('tasks', {
+                    id: randomUUID(),
+                    ...req.body
+                });
+
+                return res
                     .writeHead(201)
                     .end();
+            }else{
+                return res
+                    .writeHead(400)
+                    .end();
+            }
+            
+
+            
         }
     },
     {
@@ -36,20 +48,44 @@ export const routes = [
     {
         method: 'PUT',
         path: buildRoutePath('/tasks/:id'),
-        handler: (req, res) => {
+        handler: async (req, res) => {
             const { id } = req.params;
+
+            let tasks = await fs.readFile(databasePath, 'utf-8');
+            tasks = JSON.parse(tasks);
             
-            database.update('tasks', id, req.body);
+            const task = tasks['tasks'].findIndex(row => row.id === id);
+
+            if(task == -1) {
+                return res.writeHead(404).end(JSON.stringify({message: 'Tarefa não encontrada'}))
+            }
+
+            const { title, description } = req.body;
+
+            if(title && description) {
+                database.update('tasks', id, req.body);
             
-            return res.writeHead(204).end();
+                return res.writeHead(204).end();
+            }else{
+                return res.writeHead(400).end();
+            }
         }
     },
     {
         method: 'DELETE',
         path: buildRoutePath('/tasks/:id'),
-        handler: (req, res) => {
+        handler: async (req, res) => {
 
             const { id } = req.params;
+
+            let tasks = await fs.readFile(databasePath, 'utf-8');
+            tasks = JSON.parse(tasks);
+            
+            const task = tasks['tasks'].findIndex(row => row.id === id);
+
+            if(task == -1) {
+                return res.writeHead(404).end(JSON.stringify({message: 'Tarefa não encontrada'}))
+            }
 
             database.delete('tasks', id);
 
@@ -60,9 +96,18 @@ export const routes = [
     {
         method: 'PATCH',
         path: buildRoutePath('/tasks/:id/complete'),
-        handler: (req, res) => {
+        handler: async (req, res) => {
             const { id } = req.params;
 
+            let tasks = await fs.readFile(databasePath, 'utf-8');
+            tasks = JSON.parse(tasks);
+            
+            const task = tasks['tasks'].findIndex(row => row.id === id);
+
+            if(task == -1) {
+                return res.writeHead(404).end(JSON.stringify({message: 'Tarefa não encontrada'}))
+            }
+            
             database.completeTask('tasks', id);
 
             return res.writeHead(204).end();
